Move legacy models into legacy.py
